+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/wagner/opmodes/mechanismtests/TestArm.java
@@ -1,5 +1,6 @@
 package org.firstinspires.ftc.teamcode.wagner.opmodes.mechanismtests;
 
+import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.PoseVelocity2d;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
@@ -13,7 +14,7 @@ import org.firstinspires.ftc.teamcode.wagner.nggamepad.NGGamepad;
 public class TestArm extends LinearOpMode {
     @Override
     public void runOpMode() throws InterruptedException {
-        MecanumDrive drive = new MecanumDrive(hardwareMap, GlobalStorage.pose);
+        MecanumDrive drive = new MecanumDrive(hardwareMap, new Pose2d(0,0,0));
 
         NGGamepad ng_gamepad1 = new NGGamepad(gamepad1);
         Arm arm = new Arm();