+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/wagner/opmodes/mechanismtests/TestHanger.java
@@ -1,5 +1,6 @@
 package org.firstinspires.ftc.teamcode.wagner.opmodes.mechanismtests;
 
+import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.PoseVelocity2d;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
@@ -11,7 +12,7 @@ import org.firstinspires.ftc.teamcode.wagner.mechanisms.Hanger;
 public class TestHanger extends LinearOpMode {
     @Override
     public void runOpMode() throws InterruptedException {
-        MecanumDrive drive = new MecanumDrive(hardwareMap, GlobalStorage.pose);
+        MecanumDrive drive = new MecanumDrive(hardwareMap, new Pose2d(0,0,0));
         Hanger hanger = new Hanger();
         hanger.init(hardwareMap);
 
@@ -26,8 +27,6 @@ public class TestHanger extends LinearOpMode {
         );
 
         drive.setDrivePowers(input);
-
-        hanger.run(gamepad1, gamepad2, drive, telemetry);
         telemetry.update();
     }
 }